Chore: Move user org dto to dtos
