+++ pkg/services/sqlstore/org_users.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/grafana/grafana/pkg/api/dtos"
 	"github.com/grafana/grafana/pkg/bus"
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/util"
@@ -95,8 +96,8 @@ func (ss *SQLStore) UpdateOrgUser(ctx context.Context, cmd *models.UpdateOrgUser
 	})
 }
 
-func (ss *SQLStore) GetOrgUsers(ctx context.Context, query *models.GetOrgUsersQuery) error {
-	query.Result = make([]*models.OrgUserDTO, 0)
+func (ss *SQLStore) GetOrgUsers(ctx context.Context, query *dtos.GetOrgUsersQuery) error {
+	query.Result = make([]*dtos.OrgUserDTO, 0)
 
 	sess := x.Table("org_user")
 	sess.Join("INNER", x.Dialect().Quote("user"), fmt.Sprintf("org_user.user_id=%s.id", x.Dialect().Quote("user")))
@@ -147,9 +148,9 @@ func (ss *SQLStore) GetOrgUsers(ctx context.Context, query *models.GetOrgUsersQu
 	return nil
 }
 
-func (ss *SQLStore) SearchOrgUsers(ctx context.Context, query *models.SearchOrgUsersQuery) error {
-	query.Result = models.SearchOrgUsersQueryResult{
-		OrgUsers: make([]*models.OrgUserDTO, 0),
+func (ss *SQLStore) SearchOrgUsers(ctx context.Context, query *dtos.SearchOrgUsersQuery) error {
+	query.Result = dtos.SearchOrgUsersQueryResult{
+		OrgUsers: make([]*dtos.OrgUserDTO, 0),
 	}
 
 	sess := x.Table("org_user")
